Melee weapon swing animation replaced with jab animation for cleaner pixel scaling

+++ Utilities/NumberTools.cs
@@ -4,11 +4,11 @@ namespace Quest.Utilities;
 
 public static class NumberTools
 {
-    public static int RoundTo<T>(T num, float multiple) where T : INumber<T>
+    public static float RoundTo<T>(T num, float multiple) where T : INumber<T>
     {
         if (multiple <= 0 || num == T.Zero)
             return 0;
-        return (int)(MathF.Round(float.CreateChecked(num) / multiple) * multiple);
+        return (float)(MathF.Round(float.CreateChecked(num) / multiple) * multiple);
     }
     public static void CycleDown(ref int num, int max)
     {
